Dynamic

Description

Documentation

data Dynamic :: Type where

Dynamically typed values. Values must be monomorphic. Attempting to create a Dynamic from a polymorphic value will result an an ambiguous type variable.

>>> x = Dynamic (3 :: Int)
>>> :{
case x of
  Dynamic (y :: String) -> length y
  Dynamic () -> 0
  Dynamic (x :: Int) -> x * 2
  _ -> 42
:}
6

Constructors

MkDynamic :: TypeRep a -> a -> Dynamic

Bundled Patterns

pattern Dynamic :: forall a. Typeable a => a -> Dynamic

($@) infixl 1

Arguments

:: (IsDynamic a, IsDynamic b, MonadThrow m)

=> a
A function of type x -> y

-> b
A compatible value of type x

-> m Dynamic
The resulting value of type y

Dynamically typed function application. The first argument must be a function, and the second argument a compatible value, else a DynamicTypeError will be thrown.

>>> :{
case ((+ 3) :: Int -> Int) $@ (4 :: Int) of
  Just (Dynamic (x :: Int)) -> Just x
  _ -> Nothing
:}
Just 7

>>> x = Dynamic (4 :: Int)
>>> f = Dynamic ((+ 3) :: Int -> Int)
>>> :{
case f $@ x of
  Just (Dynamic (x :: Int)) -> Just x
  _ -> Nothing
:}
Just 7

(*@) :: (IsDynamic a, IsDynamic b, MonadThrow m) => m a -> b -> m Dynamic infixl 1

This operator is somewhat analogous to <*>. It is useful for chaining function application.

>>> x = Dynamic (4 :: Int)
>>> y = Dynamic (3 :: Int)
>>> f = Dynamic ((+) @Int)
>>> :{
case f $@ x *@ y of
  Just (Dynamic (x :: Int)) -> Just x
  _ -> Nothing
:}
Just 7

(.@) infixr 8

Arguments

:: (IsDynamic a, IsDynamic b, MonadThrow m)

=> a
A function of type y -> z

-> b
A function of type x -> y

-> m Dynamic
A function of type x -> z

Dynamic function composition.

>>> f = Dynamic ((+ 3) :: Int -> Int)
>>> g = Dynamic ((* 2) :: Int -> Int)
>>> :{
case f .@ g *@ (4 :: Int) of
  Just (Dynamic (x :: Int)) -> Just x
  _ -> Nothing
:}
Just 11

fromDynamic :: forall a m. (MonadThrow m, Typeable a) => Dynamic -> m a

Extract a dynamic value to a known concrete type.

>>> x = Dynamic (4 :: Int)
>>> fromDynamic @Int @Maybe x
Just 4

>>> fromDynamic @String @Maybe x
Nothing

fromDynamicOr :: Typeable a => a -> Dynamic -> a

Extract a dynamic value to a known concrete type with a default in case of failure.

>>> x = Dynamic (4 :: Int)
>>> fromDynamicOr "Hello" x
"Hello"

>>> fromDynamicOr (7 :: Int) x
4

dynamicRep :: Dynamic -> SomeTypeRep

Query the TypeRep of the dynamically typed value.

class IsDynamic a where

A class to allow implicit demotion of statically typed values to dynamically typed ones.

Methods

toDynamic :: a -> Dynamic

data DynamicTypeError :: Type where

A runtime exception generated by dynamic typechecking.

Constructors

DynamicApplication

Fields

:: TypeRep a

The actual type of the function

-> TypeRep b

The actual type of the argument

-> DynamicTypeError

DynamicExtraction

Fields

:: TypeRep a

The actual type of the value

-> TypeRep b

The type attempted to extract to

-> DynamicTypeError

DynamicComposition

Fields

:: TypeRep a

The actual type of the left function

-> TypeRep b

The actual type of the right function

-> DynamicTypeError